+++ moCOMOco/apps/posts/models/application.py
@@ -1,6 +1,6 @@
 from django.db import models
 from django.conf import settings
-from apps import Post
+from apps.posts.models.post import Post
 
 ROLE_CHOICES = [
     ('designer', '디자이너'),
